FIX: Switch content-length check to long

+++ wayback-core/src/main/java/org/archive/wayback/replay/TransparentReplayRenderer.java
@@ -48,7 +48,7 @@ public class TransparentReplayRenderer implements ReplayRenderer {
 	
 	// TODO: Figure out best way to generalize this, but probably good default
 	// Add special don't cache header in case of at least 100M
-	private final static int NOCACHE_THRESHOLD = 100000000;
+	private final static long NOCACHE_THRESHOLD = 100000000L;
 
 	private final static String NOCACHE_HEADER_NAME = "X-Accel-Buffering";
 	private final static String NOCACHE_HEADER_VALUE = "no";
@@ -88,10 +88,10 @@ public class TransparentReplayRenderer implements ReplayRenderer {
 		if(origLength != null) {
 			headers.put(HttpHeaderOperation.HTTP_LENGTH_HEADER, origLength);
 			
-			int contentLength = -1;
+			long contentLength = -1;
 			
 			try {
-			    contentLength = Integer.parseInt(origLength);
+			    contentLength = Long.parseLong(origLength);
 			} catch (NumberFormatException n) {
 			    
 			}
